feat: 애플리케이션이 FirebaseMessaging 인프라스트럭처에 직접 의존하고 있는 문제 개선
