+++ src/main/java/com/matsinger/barofishserver/domain/notification/application/NotificationQueryService.java
@@ -2,7 +2,7 @@ package com.matsinger.barofishserver.domain.notification.application;
 
 import com.matsinger.barofishserver.domain.notification.domain.Notification;
 import com.matsinger.barofishserver.domain.notification.repository.NotificationRepository;
-import com.matsinger.barofishserver.utils.fcm.FcmService;
+import com.matsinger.barofishserver.utils.fcm.FcmServiceProd;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class NotificationQueryService {
     private final NotificationRepository notificationRepository;
-    private final FcmService fcmService;
+    private final FcmServiceProd fcmService;
     public Page<Notification> selectNotificationListWithUserId(Integer userId, PageRequest pageRequest) {
         return notificationRepository.findAllByUserId(userId, pageRequest);
     }
