+++ src/main/java/com/matsinger/barofishserver/utils/fcm/config/DummyFcmConfig.java
@@ -1,0 +1,16 @@
+package com.matsinger.barofishserver.utils.fcm.config;
+
+import com.google.firebase.messaging.FirebaseMessaging;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+
+@Configuration
+@Profile("local")
+public class DummyFcmConfig {
+
+    @Bean
+    public FirebaseMessaging firebaseMessaging() {
+        return null;
+    }
+}
